+++ test/existence_throw_test.exs
@@ -22,8 +22,16 @@ defmodule Existence.ThrowTest do
       assert :error == Existence.get_state()
     end
 
+    test "get_state!/0" do
+      assert :error == Existence.get_state!()
+    end
+
     test "get_checks/0" do
       assert [check: {{:nocatch, :check_throw}, _stacktrace}] = Existence.get_checks()
+    end
+
+    test "get_checks!/0" do
+      assert [check: {{:nocatch, :check_throw}, _stacktrace}] = Existence.get_checks!()
     end
   end
 
@@ -45,8 +53,16 @@ defmodule Existence.ThrowTest do
       assert :error == Existence.get_state()
     end
 
+    test "get_state!/0" do
+      assert :error == Existence.get_state!()
+    end
+
     test "get_checks/0" do
       assert [check: {{:nocatch, :check_throw}, _stacktrace}] = Existence.get_checks()
+    end
+
+    test "get_checks!/0" do
+      assert [check: {{:nocatch, :check_throw}, _stacktrace}] = Existence.get_checks!()
     end
   end
 
@@ -68,8 +84,16 @@ defmodule Existence.ThrowTest do
       assert :error == Existence.get_state()
     end
 
+    test "get_state!/0" do
+      assert :error == Existence.get_state!()
+    end
+
     test "get_checks/0" do
       assert [check: {{:nocatch, :check_throw}, _stacktrace}] = Existence.get_checks()
+    end
+
+    test "get_checks!/0" do
+      assert [check: {{:nocatch, :check_throw}, _stacktrace}] = Existence.get_checks!()
     end
   end
 
@@ -91,8 +115,16 @@ defmodule Existence.ThrowTest do
       assert :error == Existence.get_state()
     end
 
+    test "get_state!/0" do
+      assert :error == Existence.get_state!()
+    end
+
     test "get_checks/0" do
       assert [check: {{:nocatch, :check_throw}, _stacktrace}] = Existence.get_checks()
+    end
+
+    test "get_checks!/0" do
+      assert [check: {{:nocatch, :check_throw}, _stacktrace}] = Existence.get_checks!()
     end
   end
 
@@ -115,8 +147,16 @@ defmodule Existence.ThrowTest do
       assert :error == Existence.get_state()
     end
 
+    test "get_state!/0" do
+      assert :error == Existence.get_state!()
+    end
+
     test "get_checks/0" do
       assert [check: {{:nocatch, :check_throw}, _stacktrace}] = Existence.get_checks()
+    end
+
+    test "get_checks!/0" do
+      assert [check: {{:nocatch, :check_throw}, _stacktrace}] = Existence.get_checks!()
     end
   end
 
@@ -139,8 +179,16 @@ defmodule Existence.ThrowTest do
       assert :error == Existence.get_state()
     end
 
+    test "get_state!/0" do
+      assert :error == Existence.get_state!()
+    end
+
     test "get_checks/0" do
       assert [check: {{:nocatch, :check_throw}, _stacktrace}] = Existence.get_checks()
+    end
+
+    test "get_checks!/0" do
+      assert [check: {{:nocatch, :check_throw}, _stacktrace}] = Existence.get_checks!()
     end
   end
 
@@ -162,8 +210,16 @@ defmodule Existence.ThrowTest do
       assert :error == Existence.get_state()
     end
 
+    test "get_state!/0" do
+      assert :error == Existence.get_state!()
+    end
+
     test "get_checks/0" do
       assert [check: {{:nocatch, :check_throw}, _stacktrace}] = Existence.get_checks()
+    end
+
+    test "get_checks!/0" do
+      assert [check: {{:nocatch, :check_throw}, _stacktrace}] = Existence.get_checks!()
     end
   end
 
@@ -186,8 +242,16 @@ defmodule Existence.ThrowTest do
       assert :error == Existence.get_state()
     end
 
+    test "get_state!/0" do
+      assert :error == Existence.get_state!()
+    end
+
     test "get_checks/0" do
       assert [check: {{:nocatch, :check_throw}, _stacktrace}] = Existence.get_checks()
+    end
+
+    test "get_checks!/0" do
+      assert [check: {{:nocatch, :check_throw}, _stacktrace}] = Existence.get_checks!()
     end
   end
 
@@ -210,8 +274,16 @@ defmodule Existence.ThrowTest do
       assert :error == Existence.get_state()
     end
 
+    test "get_state!/0" do
+      assert :error == Existence.get_state!()
+    end
+
     test "get_checks/0" do
       assert [check: {{:nocatch, :check_throw}, _stacktrace}] = Existence.get_checks()
+    end
+
+    test "get_checks!/0" do
+      assert [check: {{:nocatch, :check_throw}, _stacktrace}] = Existence.get_checks!()
     end
   end
 end