+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## v0.2.0 unreleased
+
+### Enhancements
+* [Existence] Added `get_state!/1` and `get_checks!/1`.
+* [Existence] Possible breaking change: current `get_state/1` and `get_checks/1` do not raise if
+  ETS table doesn't exist. They return `:undefined` instead.
+* [Existence.Plug] Added `:raising?` plug option.
+
 ## v0.1.1 [2022-03-23]
 
 Available on hex.pm.